requirements.txt
opencv_python==4.4.0.44
torch==1.7.0
matplotlib==3.3.2
h5py==2.10.0
torchvision==0.8.1
tqdm==4.50.2
torchsummary==1.5.1
numpy==1.19.2
Pillow==8.3.1
scikit_learn==0.24.2
